Don't allow environment variables to override critical build flags
These flags can still be overriden in the invocation of `make`, i.e.
`make ENV=val`, but that should be done with care, as these flags
are likely to control things like the layout of structs, which should
be the same in the library and the core, otherwise we will get
hard-to-understand errors.