Remove unnecessary flags in Legion CMake build (#932)
* Remove unnecessary flags in Legion CMake build.

The flag --root here is a workaround for buggy behavior in `pip install`
where pip would ignore the --prefix setting when --global-option was
supplied. As it turns out, --global-option is deprecated anyway so we
removed in Legion. As a result, this is no longer required; the prefix
is already set in the upstream project and the need for --root goes away
with the removal of --global-option.

See this Legion MR for details:

https://gitlab.com/StanfordLegion/legion/-/merge_requests/1072

* Make sure to use the latest Legion

---------

Co-authored-by: Manolis Papadakis <manopapad@gmail.com>
Co-authored-by: Manolis Papadakis <mpapadakis@nvidia.com>